add snake2.py
A slightly more complicated version of the existing snake.py, with a "blueberry" that will make you go faster for a short time but increase the snake's length by 5, as well as alternate controls (WASD), and the ability to quit by pressing "q". I also added a "rock" and "thorns", which will restart the game if bumped itno.